tala/requirements.txt
Django==1.11.3
django-adminlte-admin==0.4.11
django-adminlte2==0.1.6
django-chartjs==1.2
djangorestframework==3.6.4
future==0.16.0
Markdown==2.6.8
pexpect==4.2.1
ptyprocess==0.5.2
pyipmi==0.11.0
python-ipmi===fatal-.Not.a.git.repository.-or.any.of.the.parent.directories-..git
pytz==2017.2
six==1.10.0
